requirements.txt
certifi==2024.6.2
fastapi==0.111.0
matplotlib==3.9.0
numpy==1.26.4
Pillow==10.3.0
pydantic==2.7.3
pymongo==4.7.3
Requests==2.32.3
scikit-image==0.22.0